Fix the of the misplaced Teacher.GetById handler inside the Module UseCase
